requirements.txt
torch==2.3.1
torchvision==0.18.1
pillow==11.0.0
matplotlib==3.9.2